client: Kick cap update if max size requested
If the MDS restarts before a max size update makes
it to the journal, the client hangs waiting for the
cap_grant with max size update.  This patch adds the
cap to a list when a max size increase is requested, so
that if the mds gets restarted, we can kick the those
cap requests and get back the cap grant.

Fixes #4582.
Signed-off-by: Sam Lang <sam.lang@inktank.com>